Handle already paused
